+++ cluster/client.py
@@ -1,6 +1,6 @@
 import argparse
 
-from cluster.cluster import Cluster
+from cluster import cluster
 
 
 def main():
@@ -21,9 +21,46 @@ def main():
         '-a', '--all', action='store_true',
         help='Display all checks (any states)'
     )
+    parser_deploy = subparsers.add_parser(
+        'deploy', help='Deploy or re-deploy a service'
+    )
+    parser_deploy.add_argument(
+        'repo',
+        help='The repo name or whole form ('
+             'ssh://git@git.example.com:22/project-slug/repo-name) '
+             'for new service.'
+    )
+    parser_deploy.add_argument(
+        'branch',
+        help='The branch to deploy'
+    )
+    parser_deploy.add_argument(
+        '--master',
+        metavar='NODE',
+        help='Node where to deploy the master (required for new service)'
+    )
+    parser_deploy.add_argument(
+        '--slave',
+        metavar='NODE',
+        help='Slave node'
+    )
+
+    parser_deploy.add_argument(
+        '-w', '--wait',
+        action='store_true',
+        help='Wait the end of deployment before stop the script. Raise an'
+             'exception if deployment failed in the given time'
+    )
+    parser_deploy.add_argument(
+        '-t', '--timeout',
+        type=int,
+        default=cluster.DEFAULT_TIMEOUT,
+        help='Time in second to let a chance to deploy the service before'
+             'raising an exception (ignored without ``--wait`` option)'
+    )
 
     def init(args):
-        return Cluster(
+        return cluster.Cluster(
             args.consul
         )
 
@@ -36,7 +73,19 @@ def main():
                 for name, status, _ in service['checks']:
                     print("    - Cehck ({}): {}".format(status, name))
 
+    def cluster_deploy(args):
+        cluster = init(args)
+        cluster.deploy(
+            args.repo,
+            args.branch,
+            master=args.master,
+            slave=args.slave,
+            wait=args.wait,
+            timeout=args.timeout
+        )
+
     parser_checks.set_defaults(func=cluster_checks)
+    parser_deploy.set_defaults(func=cluster_deploy)
 
     args = parser.parse_args()
     if hasattr(args, 'func'):