+++ CHANGELOG.rst
@@ -4,4 +4,6 @@ Change log
 Unreleased
 ----------
 
+* Implement deploy subcommand to Deploy or re-deploy a service
+
 * First implementation, list checks per nodes